feat: add split_to_appearing_words() operator
split_to_appearing_words() is a new VSE operator that takes a text strip for
multiple words and splits it into individual text strips. These created strips
are offset in both time (so they 'appear' as playback progresses) and space (so
they appear in the same place they would as if they were in the original text
strip.

Customisations of time offset (etc) are not possible yet.

This commit also removes the debug text info operator.